Reduce number of rows on server console
It was found that under an Arch Linux setup with i3 as a window manager
that the server console would shrink to a super tiny size. (To see how
this looks, try resizing the window such that it is smaller than the
console text area). The reason for this is that when a window or
container is smaller than a text area, it disregards its preferred size
and snaps to a much smaller size. This bug manifests itself rarely and
only on specific setups, however just to be safe, the number of rows was
reduced.